kobe2[lte,wifi]: Add ability to reboot into brom mode
The only way to boot into bootrom mode on these devices is by opening them
and shorting the required test points, which is rather tedious and impractical.

To eliminate this inconvenience, we add two new software methods for rebooting
into bootrom mode. The first method is a fastboot command 'oem reboot-emergency'
which, when executed, will instantly reboot the device into bootrom mode.

The second method forces bootrom mode when the boot mode is FACTORYM, so bootrom
reboot can be achieved by communicating with the Preloader directly, giving us an
alternative when fastboot mode is unavailable.

Also, while we're at it, fix invalid lock state offsets for kobe2lte, which were
not affecting functionality but made fastboot oem lock-state report the device
as locked when it was actually unlocked.